Add files via upload
Created an inventory management database to control the stocks and orders of an electronics store. it has different tables for orders, customers, stock, products and suppliers.
i have also created stored procedures for insertion of new record in these table and  view to retrieve the bill of a particular purchase.